Move ServiceInfo to types package
This commit moves definition of ServiceInfo to package types. It
also modifies config.go so that the function LoadServiceInfo now
returns a pointer to a struct of this type instead of returning
service name, description and data individually. This change is
for future compatibility with tools for creating language bingings.
